moved select test cases out of managed state
