+++ src/__tests__/Async.test.js
@@ -44,7 +44,7 @@ cases('load option prop with defaultOptions true', ({ props, expectOptionLength 
  * confirmed by logging in component that loadOptions is resolved and options are available
  * but still loading options is rendered
  */
-cases('load options props with', ({ props, expectloadOptionsLength }) => {
+cases('load options props with no default options', ({ props, expectloadOptionsLength }) => {
   let asyncSelectWrapper = mount(<Async {...props} />);
   let inputValueWrapper = asyncSelectWrapper.find('div.react-select__input input');
   asyncSelectWrapper.setProps({ inputValue: 'a' });
